backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
yfinance>=0.2.40
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
apscheduler==3.10.4
pyyaml>=6.0.1
redis>=5.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
aiohttp>=3.9.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-auth>=2.25.0
